+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 **Next release**
 
 * Fixed setting dangerous restrictions ([issue](https://github.com/M66B/XPrivacy/issues/876))
+* Fixed web view usage data
 * Attempt to fix user agent restriction for KitKat ([issue](https://github.com/M66B/XPrivacy/issues/825))
 * Show relative usage time
 * Filter application details on data usage